Extract Runtime and ExecutionContext concept from Executor
This is a precursor to some optimisation work which
requires some of the current Executor's internals.

This commit splits out 2 classes:

- Runtime which implements helpers around the resolver's runtime
(AsyncIO, threads, etc.). Outside of BlockingExecutor this should the
class most consumer interact with when customising the execution.
- ExecutionContext which extracts most of the caches and helper methods
used to traverse the GraphQL during execution.

This also includes related updates to the ResolveInfo class, examples
and internal utility functions.