Schema directives must now specify their defiinition.
Separating definition and implementation lead to duplication of
information which could be problematic in large code bases.

This change enforces explicit definition of the directive on
SchemaDirective subclasses as either:

    - A Directive instance.
    - A directive name (string) which explicitely delegates to the
    schema at application time.

This still allows defining schema directives in SDL documents when
appropriate or simply while defaulting to a a single location case when
defining the Directive on the SchemaDirective subclasses.

Following from this, build_schema now accepts directives as a Sequence
instead of a Mapping (breaking) and applying schema directives will
modify the schema by adding the new directive definitions.