Ignore unknown directives in UniqueDirectivesPerLocationChecker rule.
The implementation was treating all directives as known and validating.
This could be confusing when bailing out for repeatable
directives as we wouldn't know whether an unknown directive
should be treated as repeatable or not.

This changes properly raises SkipNode in KnownDirectivesChecker
and lets UniqueDirectivesPerLocationChecker ignore them as they
should have been triggered an error message already.