Minor cleanup / optimization of field resolution.
The field execution layer is the most costly in production, this commit
is a mixed bag of optimizations.

The most notable is dropping a bunch of `cast` calls in the runtime
implementations in favor of `# type: ignore` comments. I think this is a
valid trade off here as cast, while still minimal, has a runtime cost
which adds up quickly on large queries.