Unify AST printing code.
So far, there was ASTprinter which could print arbitrary ASTs and
ASTSchemaPrinter which was specialized for printing Schema Definition
Language documents.

Rather than maintaining 2 distinct ways of printing documents this
commit  replaces this with a 2 step process to print a schema: 1)
convert the schema to an AST with ASTSchemaConverter and 2) print the
resulting AST with ASTPrinter.

Notes:

- The previous implementation had some custom handling for descriptions
(e.g. wrap to 120 characters, keep on 1 line if short enough). This is no
gone as it simplified the code and the resulting documents are
semantically equivalent.
- Some bugs, omissions have been fixed in ASTPrinter.
- This is likely slower than before, however printing a schema is
unlikely to be in the hot path so this should be fine.