Always validate queries according to the spec.
Through a couple of iterations, we'd previously settled on replacing the
specified rules when using custom validators with the `graphql` entry
point and its derivatives. This meant users had to re-specify the
default validator, which could be easily missed.

This change makes it so the default rules are always included and cannot
be disabled. It also replaces the `validators` parameter with an
`additional_validators` parameter for clarity.