Fix handling of recursive input types in build_schema()
Recursive input types would lead to recursion errors when defined in
their SDL form through build_schema().

The fix is quite involved and requires correctly making the input
fields' types lazy while building the types (as we do for object
fields' types), but also the default value given that to parse
and validate a default value for complex types we need the type
object.

Making the default value lazy highlighted an issue with lazy values in
the type builder: as evaluation would be delayed until first access,
this could well be at runtime if the value was not used during the rest
of the SDL building process; which is exactly the case for default values.
To solve this, the ASTTypeBuilder now maintains a list of lazy objects
that were created and we _flush_ that list once all the types have been
built. This ensure any lazy value that was not required through other
types gets called before we return the schema to the user so there is no
surprise at runtime. The main caveat is that when an exception is raised
in one of these delayed calls, the callsatck will be somewhat removed
from the code where it actually came from and point to the line where we
flush.